django/mgnify_bgcs/services/umap_train/requirements.txt
numpy
pandas
umap-learn
scikit-learn
joblib
pyarrow
